requirements.txt
aiogram==3.19.0
python-dotenv==1.1.0
requests-2.32.3
requests~=2.32.3